feat(task2): add rational numbers support

+++ task2/task2.py
@@ -1,4 +1,5 @@
 import sys
+from decimal import Decimal
 
 
 def main():
@@ -6,14 +7,14 @@ def main():
     points_file_path = sys.argv[2]
 
     with open(coords_file_path, mode="r", encoding="utf8") as f:
-        center_x, center_y = [int(v) for v in f.readline().split()]
-        radius_x, radius_y = [int(v) for v in f.readline().split()]
+        center_x, center_y = [Decimal(v) for v in f.readline().split()]
+        radius_x, radius_y = [Decimal(v) for v in f.readline().split()]
 
     with open(points_file_path, mode="r", encoding="utf8") as f:
         for point in f:
             x, y = point.split()
 
-            res = (int(x) - center_x)**2 / radius_x**2 + (int(y) - center_y)**2 / radius_y**2
+            res = (Decimal(x) - center_x)**2 / radius_x**2 + (Decimal(y) - center_y)**2 / radius_y**2
 
             if res > 1:
                 # снаружи
